+++ src/main/java/shop/wannab/couponservice/repository/CouponRepository.java
@@ -1,8 +1,8 @@
 package shop.wannab.couponservice.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.wannab.couponservice.domain.Coupon;
-import shop.wannab.couponservice.domain.CouponPolicy;
+import shop.wannab.couponservice.domain.coupon.Coupon;
+import shop.wannab.couponservice.domain.couponpolicy.CouponPolicy;
 
 public interface CouponRepository extends JpaRepository<Coupon, Long> {
     Coupon findByUserIdAndCouponPolicy(Long userId, CouponPolicy couponPolicy);
