+++ src/main/java/shop/wannab/couponservice/repository/PolicyTargetBookRepository.java
@@ -2,8 +2,8 @@ package shop.wannab.couponservice.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.wannab.couponservice.domain.CouponPolicy;
-import shop.wannab.couponservice.domain.PolicyTargetBook;
+import shop.wannab.couponservice.domain.couponpolicy.CouponPolicy;
+import shop.wannab.couponservice.domain.couponpolicy.PolicyTargetBook;
 
 public interface PolicyTargetBookRepository extends JpaRepository<PolicyTargetBook, Long> {
     Optional<PolicyTargetBook> findByCouponPolicy(CouponPolicy couponPolicy);
