+++ src/main/java/shop/wannab/couponservice/service/CouponService.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponentsBuilder;
-import shop.wannab.couponservice.domain.Coupon;
-import shop.wannab.couponservice.domain.CouponPolicy;
+import shop.wannab.couponservice.domain.coupon.Coupon;
+import shop.wannab.couponservice.domain.couponpolicy.CouponPolicy;
 import shop.wannab.couponservice.domain.enums.CouponType;
 import shop.wannab.couponservice.repository.CouponPolicyRepository;
 import shop.wannab.couponservice.repository.CouponRepository;
